Refactor Vulkan pooling kernels to support batched input and improve numerical stability
- Updated Vulkan compute shaders for pooling operations (bf16, f16, f64) to handle batched inputs by decoding the z dimension into batch and channel indices.
- Enhanced the handling of input and output indices to ensure correct memory access patterns for batched data.
- Introduced Taylor series approximation for sine function in special_math_unary.comp to improve precision for inputs near multiples of π.
- Modified Vulkan kernel registration to reflect changes in LSTMForward, ensuring improved numerical stability with new sigmoid/tanh formulations.
- Adjusted memory operations in Vulkan backend to correctly handle float conversions and prevent silent errors during tensor fills.
- Improved numerical stability tests for various operations, including log_softmax and FP16 saturation, to ensure consistent behavior across backends.
- Updated tests to account for new tolerances in LSTM and GRU multi-layer operations due to accumulated numerical drift.

+++ src/backends/vulkan/vulkan_ops_fft.cpp
@@ -1910,7 +1910,7 @@ auto VulkanBackend::dispatchIRFFT(const Tensor& input, int64_t dim, int64_t n,
     // For norm=="forward": target = 1, applied = 2/N, correction = N/2
 
     if (std::abs(correction - 1.0) > 1e-15) {
-        auto scale_tensor = dispatchFull({1}, static_cast<float>(correction), real_dtype);
+        auto scale_tensor = dispatchFull({1}, correction, real_dtype);
         output = dispatchBinaryOp("mul", output,
             dispatchExpand(scale_tensor, std::vector<int64_t>(out_shape.begin(), out_shape.end())));
     }
